replacing gca and gcf with double(gca) and double(gcf) respectively in the whole repository to match new syntax

+++ ExperPort/Analysis/@Analyzer/analyze_cp_history.m
@@ -141,7 +141,7 @@ if ~isempty(h2) & length(u) > 1,
     xlabel(h2, 'Percentile')
     ylabel(h2, 'CPoke duration (s)')
     gridpts = [0 25 50 75 95]; % must always contain 0
-    set(gca, 'XTick', gridpts, 'XGrid', 'on', 'Xlim', gridpts([1 end]));
+    set(double(gca), 'XTick', gridpts, 'XGrid', 'on', 'Xlim', gridpts([1 end]));
 
     p = zeros(size(gridpts)); p(1) = 1; empty_flag = 0;
     for i=2:length(gridpts),
@@ -151,7 +151,7 @@ if ~isempty(h2) & length(u) > 1,
         end;
     end;
     if ~empty_flag,
-        if min(diff(x(p)))>0, set(gca, 'YTick', x(p), 'Ygrid', 'on', 'YLim', x(p([1 end]))); end;
+        if min(diff(x(p)))>0, set(double(gca), 'YTick', x(p), 'Ygrid', 'on', 'YLim', x(p([1 end]))); end;
     end;
