+++ ExperPort/Analysis/ProAnti/lesion_rats.m
@@ -12,7 +12,7 @@ pa_s={'anti','pro'};
 m={'x' 'o' 's' '.' 'x' 'o' 's' '.'};
 for pax=1:2
 	subplot(1,2,pax);
-	set(gca,'FontSize',14);
+	set(double(gca),'FontSize',14);
 	hold on;
 	for rx=1:numel(all_rats)
 		[perf]=bdata('select total_correct from sessions where ratname="{S}" and comments="{S}" and sessiondate>"2007-11-23" order by sessiondate', all_rats{rx}, pa_s{pax});
